+++ packages/demo-next/pages/blocks.tsx
@@ -26,17 +26,15 @@ import {
   InlineTextField,
   InlineBlocks,
   BlocksControls,
-  BlockText,
-  BlockImage,
-  BlockTextarea,
   useInlineForm,
+  InlineTextareaField,
 } from 'react-tinacms-inline'
 
 /**
  * This is an example page that uses Blocks from Json
  */
 export default function BlocksExample({ jsonFile }) {
-  const [, form] = useJsonForm(jsonFile)
+  const [data, form] = useJsonForm(jsonFile)
 
   return (
     <ModalProvider>
@@ -47,9 +45,12 @@ export default function BlocksExample({ jsonFile }) {
           <InlineTextField name="title" />
           <InlineImageField
             name="hero_image"
+            previewSrc={formValues => formValues.hero_image}
             parse={filename => `/images/${filename}`}
             uploadDir={() => '/public/images/'}
-          />
+          >
+            {props => <ChildImage src={data.hero_image} {...props} />}
+          </InlineImageField>
         </h1>
         <Wrap>
           <InlineBlocks name="blocks" blocks={PAGE_BUILDER_BLOCKS} />
@@ -57,6 +58,10 @@ export default function BlocksExample({ jsonFile }) {
       </InlineForm>
     </ModalProvider>
   )
+}
+
+function ChildImage(props) {
+  return <img src={props.previewSrc || props.src} />
 }
 
 /**
@@ -114,13 +119,14 @@ function HeroBlock({ index }) {
 function ImageBlock({ index, data }) {
   return (
     <BlocksControls index={index}>
-      <BlockImage
+      <InlineImageField
         name="src"
+        previewSrc={formValues => {
+          return formValues.blocks[index].src
+        }}
         parse={filename => `/images/${filename}`}
         uploadDir={() => '/public/images/'}
-      >
-        <img src={data.src} alt={data.alt} />
-      </BlockImage>
+      />
     </BlocksControls>
   )
 }
@@ -139,7 +145,7 @@ const image_template: BlockTemplate = {
 
 // Testing the block styled component override
 
-const StyledBlockText = styled(BlockText)`
+const StyledBlockText = styled(InlineTextareaField)`
   color: green;
 `