+++ packages/react-tinacms-editor/src/plugins/Blockquote/index.ts
@@ -16,4 +16,4 @@ limitations under the License.
 
 */
 
-export { Menu } from './Menu'
+export * from './Menu'