+++ packages/react-tinacms-inline/src/blocks/inline-block-image.tsx
@@ -17,47 +17,14 @@ limitations under the License.
 */
 
 import * as React from 'react'
-import { BlockField } from './inline-block-field'
-import { InlineImageProps, ImageUpload } from '../inline-field-image'
-import { useCMS } from 'tinacms'
+import {
+  InlineImageProps,
+  InlineImageField,
+} from '../fields/inline-image-field'
 
-export function BlockImage({
-  name,
-  children,
-  parse,
-  uploadDir,
-}: InlineImageProps) {
-  const cms = useCMS()
-  return (
-    <BlockField name={name}>
-      {({ input, status, form }) => {
-        if (status === 'active') {
-          return (
-            <ImageUpload
-              value={input.value}
-              onDrop={async ([file]: File[]) => {
-                const directory = uploadDir(form)
-                const [media] = await cms.media.store.persist([
-                  {
-                    directory,
-                    file,
-                  },
-                ])
-                if (media) {
-                  input.onChange(parse(media.filename))
-                } else {
-                  // TODO Handle failure
-                }
-                return null
-              }}
-              {...input}
-            >
-              {children}
-            </ImageUpload>
-          )
-        }
-        return children ? children : <img src={input.value} />
-      }}
-    </BlockField>
-  )
+/**
+ * @deprecated
+ */
+export function BlockImage(props: InlineImageProps) {
+  return <InlineImageField {...props} focusRing={false} />
 }