+++ packages/react-tinacms-inline/src/fields/inline-textarea-field.tsx
@@ -16,27 +16,29 @@ limitations under the License.
 
 */
 
-import * as React from 'react'
+import React from 'react'
+import { InlineField } from '../inline-field'
 import styled from 'styled-components'
-import { InlineField } from './inline-field'
-import { InputFocusWrapper } from './styles'
+import TextareaAutosize from 'react-textarea-autosize'
+import { InlineTextFieldProps } from './inline-text-field'
+import { InputFocusWrapper } from '../styles'
 
-/**
- * InlineTextField
- */
-export interface InlineTextFieldProps {
-  name: string
-  className?: string
-}
-
-export function InlineTextField({ name, className }: InlineTextFieldProps) {
+export function InlineTextareaField({
+  name,
+  className,
+  focusRing,
+}: InlineTextFieldProps) {
   return (
     <InlineField name={name}>
       {({ input, status }) => {
         if (status === 'active') {
+          if (!focusRing) {
+            return <InlineTextarea className={className} {...input} rows={1} />
+          }
+
           return (
             <InputFocusWrapper>
-              <InlineText type="text" {...input} className={className} />
+              <InlineTextarea className={className} {...input} rows={1} />
             </InputFocusWrapper>
           )
         }
@@ -46,8 +48,9 @@ export function InlineTextField({ name, className }: InlineTextFieldProps) {
   )
 }
 
-export const InlineText = styled.input`
+const InlineTextarea = styled(TextareaAutosize)`
   width: 100%;
+  word-wrap: break-word;
   display: block;
   font-size: inherit;
   font-family: inherit;
@@ -56,6 +59,7 @@ export const InlineText = styled.input`
   color: inherit;
   letter-spacing: inherit;
   line-height: inherit;
+  margin: 0 auto;
   max-width: inherit;
   background-color: inherit;
   text-align: inherit;
@@ -65,4 +69,8 @@ export const InlineText = styled.input`
   overflow: visible;
   position: relative;
   -ms-overflow-style: none;
+
+  ::-webkit-scrollbar {
+    display: none;
+  }
 `